requirements.txt
langchain==0.3.19
langchain_community==0.3.18
langchain_core==0.3.40
langchain_openai==0.3.7
langchain_pinecone==0.2.3
langchain_text_splitters==0.3.6
langgraph==0.2.74
pinecone==6.0.1
python-dotenv==1.0.1
python_docx==1.1.2
streamlit==1.38.0
streamlit==1.42.0
